Add the function of post
On the detailed event screen allow users to share
the event through a Twitter status update (Mastodon).
The user interface contains an input text field for
the content of the post and a button to submit it.
The placeholder attribute of the text field contains
the name of the event.
If the post has been published successfully the user
interface gives feedback to the user by confirming
the post has been published. This confirmation
contains the content of the original post.

Change: EventsController.java, description.html